break apart service tests
