+++ metricbeat/module/system/service/service_test.go
@@ -20,15 +20,11 @@
 package service
 
 import (
-	"context"
-	"os"
-	"runtime"
 	"testing"
 	"time"
 
 	"github.com/coreos/go-systemd/v22/dbus"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
 	_ "github.com/elastic/beats/v7/metricbeat/module/system"
 	"github.com/elastic/elastic-agent-libs/mapstr"
@@ -46,29 +42,6 @@ var exampleUnits = []dbus.UnitStatus{
 	dbus.UnitStatus{
 		Name: "filebeat.service",
 	},
-}
-
-func TestDbusEnvConnection(t *testing.T) {
-	if runtime.GOOS != "linux" {
-		t.Skip("Test is linux-only")
-	}
-
-	// Set specific env var
-	// This format is for the newer versions of the godbus/dbus library
-	// Older versions use a format with out the `path` prefix.
-	err := os.Setenv("DBUS_SYSTEM_BUS_ADDRESS", "unix:path=/var/run/dbus/system_bus_socket")
-	require.NoError(t, err)
-
-	// call internal dbus functions
-	// This calls a lower-level bus library
-	conn, err := instrospectForUnitMethods()
-	require.NoError(t, err)
-	require.NotNil(t, conn)
-
-	// test the higher-level systemd library
-	_, err = dbus.NewWithContext(context.Background())
-	require.NoError(t, err)
-
 }
 
 func TestFormProps(t *testing.T) {
